new function system
at first only for the method selection
