+++ Stopping_distances/SD_Constants.py
@@ -1,0 +1,8 @@
+# constants
+
+
+
+# random initial velocity settings
+min_v = 0.5
+max_v = 10
+steps_v = 0.5
